[Keyboard] bm40hsrgb: move rev1 files to rev1 directory
