+++ keyboards/kprepublic/bm40hsrgb/rev1/info.json
