+++ keyboards/kprepublic/bm40hsrgb/rev1/keymaps/signynt/config.h
