+++ keyboards/kprepublic/bm40hsrgb/rev1/keymaps/signynt/rules.mk
