+++ keyboards/kprepublic/bm40hsrgb/rev1/keymaps/signynt_2_loud/rules.mk
